+++ src/libstd/user/functions/math/ceil.c
@@ -14,7 +14,7 @@ double ceil(double x)
         return NAN;
     }
 
-    if (fabs(x) > INT64_MAX)
+    if (fabs(x) > (double)INT64_MAX)
     {
         return x;
     }
